app/requirements.txt
django==4.0.4
uwsgi==2.0.20
python-dotenv==0.21.0
django-split-settings==1.1.0
psycopg2==2.9.3
django-cors-headers==3.13.0